Use proper bmap file for tests
The previous bmap file mapped the entire image in a single block. This
didn't really test if bmaps are working since there were no holes in
the map.

Created a new bmap of the etcher-test.img using `bmap create` and added
it. Updated one test for the new number of block maps and checksums.
Other tests are failing, likely because of an existing bug.