source: add imageType property
This ties in with modifications to balena-yocto-scripts and balena-img
that allows for multiple explicitly labeled image types, such as raw and
flasher, to be uploaded with a suffix.

This change adds the imageType property to BalenaS3Source, which is used
by a getter to generate the suffix to retrieve the relevant files.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>